# server/api/openapi.gen.yaml
...
components:
    parameters:
        DocListCursor:
            in: query
            name: cursor
            required: false
            schema:
                type: string
        DocListDateFrom:
            in: query
            name: dateFrom
            required: false
            schema:
                format: date
                type: string
        DocListDateTo:
            in: query
            name: dateTo
            required: false
            schema:
                format: date
                type: string
        DocListLimit:
            in: query
            name: limit
            required: false
            schema:
                default: 50
                type: integer
        DocListQ:
            in: query
            name: q
            required: false
            schema:
                type: string
        DocListStatus:
            in: query
            name: status
            required: false
            schema:
                type: string
        DocListWarehouseId:
            in: query
            name: warehouseId
            required: false
            schema:
                format: uuid
                type: string
        DocumentId:
            in: path
            name: id
            required: true
            schema:
                format: uuid
                type: string
...
    responses:
        Conflict:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Resource already exists
        Forbidden:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Authenticated but not permitted
        NotFound:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Resource not found
...
        Unauthorized:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Not authenticated
        ValidationError:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Request failed validation
    schemas:
...
        DecimalString:
            description: A decimal quantity or money value, string-encoded to avoid float loss
            example: "12.500"
            type: string
...
        DocumentActor:
            properties:
                displayName:
                    type: string
                id:
                    format: uuid
                    type: string
            required:
                - id
                - displayName
            type: object
        DocumentStatus:
            description: draft is editable; posted is immutable; reversed was cancelled by a reversal; cancelled is terminal for PO/SO (ADR-0018)
            enum:
                - draft
                - posted
                - reversed
                - cancelled
            type: string
        Error:
            properties:
                code:
                    type: string
                fields:
                    additionalProperties:
                        type: string
                    description: Per-field validation messages, when applicable
                    type: object
                message:
                    type: string
            required:
                - code
                - message
            type: object
...
        StockAdjustment:
            properties:
                createdAt:
                    format: date-time
                    type: string
                createdBy:
                    $ref: '#/components/schemas/DocumentActor'
                docDate:
                    format: date
                    type: string
                docNumber:
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                lines:
                    items:
                        $ref: '#/components/schemas/StockAdjustmentLine'
                    type: array
                notes:
                    type: string
                postedAt:
                    format: date-time
                    type: string
                postedBy:
                    $ref: '#/components/schemas/DocumentActor'
                reason:
                    type: string
                reversedById:
                    format: uuid
                    nullable: true
                    type: string
                reversesId:
                    format: uuid
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/DocumentStatus'
                warehouseId:
                    format: uuid
                    type: string
            required:
                - id
                - status
                - warehouseId
                - reason
                - docDate
                - notes
                - lines
                - createdAt
                - createdBy
            type: object
        StockAdjustmentInput:
            properties:
                docDate:
                    format: date
                    type: string
                lines:
                    items:
                        $ref: '#/components/schemas/StockAdjustmentLineInput'
                    minItems: 1
                    type: array
                notes:
                    type: string
                reason:
                    type: string
                warehouseId:
                    format: uuid
                    type: string
            required:
                - warehouseId
                - docDate
                - lines
            type: object
        StockAdjustmentLine:
            properties:
                batchId:
                    format: uuid
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                lineNo:
                    type: integer
                productId:
                    format: uuid
                    type: string
                qty:
                    $ref: '#/components/schemas/DecimalString'
                unitCost:
                    $ref: '#/components/schemas/DecimalString'
                uom:
                    type: string
            required:
                - id
                - lineNo
                - productId
                - uom
                - qty
                - unitCost
            type: object
        StockAdjustmentLineInput:
            properties:
                batchId:
                    format: uuid
                    nullable: true
                    type: string
                productId:
                    format: uuid
                    type: string
                qty:
                    allOf:
                        - $ref: '#/components/schemas/DecimalString'
                    description: Signed - positive is found stock, negative is waste/damage; nonzero
                unitCost:
                    $ref: '#/components/schemas/DecimalString'
                uom:
                    minLength: 1
                    type: string
            required:
                - productId
                - uom
                - qty
            type: object
        StockAdjustmentList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/StockAdjustment'
                    type: array
                nextCursor:
                    nullable: true
                    type: string
            required:
                - items
                - nextCursor
            type: object
...
    securitySchemes:
        sessionCookie:
            in: cookie
            name: sinta_session
            type: apiKey
info:
    description: |
        Multi-tenant inventory management for Indonesian SME distributors. Contract-first: this file is the single source of truth for both the Go server (oapi-codegen) and the TypeScript client (openapi-typescript). Conventions: camelCase fields, kebab-case plural paths, /v1 prefix (docs/reference/CONVENTIONS.md).
    title: Sinta API
    version: 0.1.0
openapi: 3.0.3
paths:
...
    /stock-adjustments:
        get:
            operationId: listStockAdjustments
            parameters:
                - $ref: '#/components/parameters/DocListStatus'
                - $ref: '#/components/parameters/DocListWarehouseId'
                - $ref: '#/components/parameters/DocListDateFrom'
                - $ref: '#/components/parameters/DocListDateTo'
                - $ref: '#/components/parameters/DocListQ'
                - $ref: '#/components/parameters/DocListCursor'
                - $ref: '#/components/parameters/DocListLimit'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockAdjustmentList'
                    description: Stock adjustments
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
            summary: List stock adjustments
            tags:
                - documents
        post:
            operationId: createStockAdjustment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StockAdjustmentInput'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockAdjustment'
                    description: Draft created
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Create a draft stock adjustment
            tags:
                - documents
    /stock-adjustments/{id}:
        delete:
            operationId: deleteStockAdjustment
            responses:
                "204":
                    description: Deleted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Delete a draft stock adjustment (rejected once posted)
            tags:
                - documents
        get:
            operationId: getStockAdjustment
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockAdjustment'
                    description: Stock adjustment
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: Get a stock adjustment
            tags:
                - documents
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        put:
            operationId: updateStockAdjustment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StockAdjustmentInput'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockAdjustment'
                    description: Draft updated
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Replace a draft stock adjustment (rejected once posted)
            tags:
                - documents
    /stock-adjustments/{id}/post:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: postStockAdjustment
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockAdjustment'
                    description: Posted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Post a draft stock adjustment (signed adjustment movements)
            tags:
                - documents
    /stock-adjustments/{id}/reverse:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: reverseStockAdjustment
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockAdjustment'
                    description: Reversal document posted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Reverse a posted stock adjustment (posts the negated movements)
            tags:
                - documents
...
security:
    - sessionCookie: []
servers:
    - url: /v1